rename announcementsmodal
