+++ bundles/framework/announcements/view/AnnouncementsPopup.jsx
