+++ bundles/framework/announcements/view/index.js
@@ -1,4 +1,4 @@
-export { showAnnouncementsPopup } from './AnnouncementsModal';
+export { showAnnouncementsPopup } from './AnnouncementsPopup';
 export { AnnouncementsCollapse } from './AnnouncementsCollapse';
 export { AnnouncementsContent } from './AnnouncementsContent';
 export { AnnouncementsHandler } from './AnnouncementsHandler';
